pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cva-cycb"
version = "0.1.0"
description = "Counterfactually-Verifiable Agents (CVA) + CyCB benchmark reproduction scaffold"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7",
]


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
